+++ src/boot.js
@@ -59,24 +59,9 @@ module.exports = async function boot (whichPackage) {
     utils.initProject(runtime)
   }
 
-  // Run Hooks
+  // Run Pre-Hooks
   if (cmdName.match(/^build:(assets|content)/)) {
-    ['prebuild', 'prebuild:production', 'prebuild:development'].forEach(async (hook) => {
-      if (runtime[hook]) {
-        const needEnv = hook.split(':')[1]
-        if (!needEnv || runtime.lanyonEnv === needEnv) {
-          let squashedHooks = runtime[hook]
-          if (_.isArray(runtime[hook])) {
-            squashedHooks = runtime[hook].join(' && ')
-          }
-          await scrolex.exe(squashedHooks, {
-            cwd : runtime.projectDir,
-            mode: 'passthru',
-          })
-          // scrolex.stick(`${hook} done`)
-        }
-      }
-    })
+    await utils.runhooks('pre', cmdName, runtime)
   }
 
   // Write all config files to cacheDir
@@ -137,7 +122,7 @@ module.exports = async function boot (whichPackage) {
     let stdio           = 'pipe'
     let mode            = 'passthru'
     if (cmdName === 'start') {
-      mode  = (process.env.SCROLEX_MODE || 'singlescroll')
+      mode = (process.env.SCROLEX_MODE || 'singlescroll')
     }
     if (cmdName === 'container:connect') {
       stdio = 'inherit'
@@ -147,12 +132,17 @@ module.exports = async function boot (whichPackage) {
     cmd = cmd.replace(/(\s|^)jekyll(\s|$)/, `$1${runtime.binDir}/jekyll$2`)
     cmd = cmd.replace(/(\s|^)bundler(\s|$)/, `$1${runtime.binDir}/bundler$2`)
 
-    await scrolex.exe(cmd, {
+    scrolex.exe(cmd, {
       cwd  : runtime.cacheDir,
       stdio: stdio,
       mode : mode,
+      fatal: true,
+    }, async (err, out) => { // eslint-disable-line handle-callback-err
+      // Run Post-Hooks
+      await utils.runhooks('post', cmdName, runtime)
     })
   } else {
     scrolex.failure(`"${cmdName}" is not a valid Lanyon command. Pick from: ${Object.keys(scripts).join(', ')}.`)
+    process.exit(1)
   }
 }